pyproject.toml
[tool.poetry]
name = "airpods"
version = "0.1.0"
description = ""
authors = ["Sergei Sorochinskii <sergei.sorochinskii@lanars.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
flake8 = "^4.0.1"
requests = "^2.26.0"
django-cors-headers = "^3.10.1"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
